requirements.txt
curl_cffi==0.7.3
playwright==1.44.0
playwright-stealth==1.0.6
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
aiofiles==23.2.1
apscheduler==3.10.4
colorlog==6.8.2
tenacity==8.3.0
